refactor: remove relative URL resolver
