+++ projects/ngx-meta/src/core/src/url-resolution/default-url-resolver.ts
@@ -1,21 +1,24 @@
+import { inject, Provider } from '@angular/core'
 import { BaseUrl } from './base-url'
-import { Provider } from '@angular/core'
-import { Router } from '@angular/router'
+import { _URL_RESOLVER, _UrlResolver } from './url-resolver'
 import { ANGULAR_ROUTER_URL } from './angular-router-url'
-import {
-  RELATIVE_URL_RESOLVER,
-  RelativeUrlResolver,
-} from './relative-url-resolver'
+import { Router } from '@angular/router'
 
-export const provideDefaultRelativeUrlResolver: (
-  baseUrl: BaseUrl,
-) => Provider = (baseUrl) => ({
-  provide: RELATIVE_URL_RESOLVER,
-  useFactory:
-    (router: Router): RelativeUrlResolver =>
-    (givenUrl) => {
-      const relativeUrl =
-        givenUrl === ANGULAR_ROUTER_URL ? router.url : givenUrl
+export const provideDefaultUrlResolver: (baseUrl: BaseUrl) => Provider = (
+  baseUrl,
+) => ({
+  provide: _URL_RESOLVER,
+  useFactory: (): _UrlResolver => {
+    const router = inject(Router)
+    return (url) => {
+      if (url === undefined || url === null) {
+        return url
+      }
+      const urlString = url.toString()
+      if (urlString.split('://').length > 1) {
+        return urlString
+      }
+      const relativeUrl = url === ANGULAR_ROUTER_URL ? router.url : urlString
       /* istanbul ignore next https://github.com/istanbuljs/istanbuljs/issues/719 */
       if (relativeUrl.length === 0) {
         return baseUrl
@@ -29,6 +32,6 @@ export const provideDefaultRelativeUrlResolver: (
             ? `/${relativeUrl}`
             : relativeUrl
       return `${baseUrl}${pathToAppend}`
-    },
-  deps: [Router],
+    }
+  },
 })
