+++ projects/ngx-meta/src/core/src/url-resolution/url-resolver.ts
@@ -1,6 +1,6 @@
-import { inject, InjectionToken } from '@angular/core'
+import { InjectionToken } from '@angular/core'
 import { AngularRouterUrl } from './angular-router-url'
-import { RELATIVE_URL_RESOLVER } from './relative-url-resolver'
+import { noOpUrlResolver } from './no-op-url-resolver'
 
 /**
  * Resolves relative URLs into absolute URLs if a base URL was provided.
@@ -12,24 +12,7 @@ import { RELATIVE_URL_RESOLVER } from './relative-url-resolver'
  */
 export const _URL_RESOLVER = new InjectionToken<_UrlResolver>(
   ngDevMode ? 'NgxMeta URL Resolver' : 'NgxMetaUR',
-  {
-    factory: () => {
-      const relativeUrlResolver = inject(RELATIVE_URL_RESOLVER)
-      return (url) => {
-        if (url === undefined || url === null) {
-          return url
-        }
-        const urlString = url.toString()
-        if (urlString.split('://').length > 1) {
-          return urlString
-        }
-        return relativeUrlResolver(
-          //👇 Due to previous `if` clause, it can't be a URL object anymore
-          url as Exclude<typeof url, URL>,
-        )
-      }
-    },
-  },
+  { factory: () => noOpUrlResolver },
 )
 
 /**
